servo setup WIP
Servos only work with manual override at the moment, and the servos do
nothing with manual override off. This will be done on Thursday. (the
following is not done yet) It will use two push button on the
controller,  and will only work in the end game using the timer.