fixed some of the bugs we were seeing(Read description)
arm encoder distance per puls has to be changed(Robot map), no clue
about the lidar, servo MIGHT be fixed, but now the value is easily
editable as a constant in the multiTool subsystem. autofinish should
also work, i think the problem was just that i was holding the button.
This branch still needs to be merged into Fresh-Start, if it works that
is